Store local state for included plugins.
Create a file containing all states for included plugin repos. When
building restore the plugin repos from this file. This creates
a reproducible which not fails on unexpected commits on remote
plugin repos.

The central piece is the source-state.sh script. Usage is basically:

   $ ./source-state-sh restore     # Clone sources and checkout commit
   $ ./source-state.sh update      # Update sources to latest HEAD
   $ ./source-state.sh save        # Update sources.state

Manual updates of plugin sources (git checkout whatever) are fine and
will be saved on next 'save' operationm.